Wire in stepper motor
This code assumes the Twiz-manager software has been set up to send only the
rotational z euler angle as a decimal (i.e. custom content of "rz_d" which
sends a value between 0.0 and 1.0). The arduino will expect the data on port
10000

The code also attempts some very crude smoothing by only updating the stepper
motor if the difference in the reading from the previous value is greater than
0.05, which greatly reduces the fidility, but results in less jittery motion.