Don't invert the reading
This inversion was added because the value coming out of the Twiz moved in the
opposite direction to the stepper driver when it is sat base-down, which means
the rotation of the stepper would move opposite to the Twiz. But in the show,
the stepper is hung upside down, meaning that we don't need to negate the value
we send to the stepper. In either case, this sort of compensation should be done
outside of this library code.